Moved coordinate and cube padding utilities into module separate from neighbourhood processing.  No functional changes.  All unit tests pass.
